+++ src/pools/pool.ts
@@ -42,9 +42,10 @@ export class PoolEmitter extends EventEmitter {}
  * Enumeration of pool events.
  */
 export const PoolEvents = Object.freeze({
-  full: 'full',
   ready: 'ready',
   busy: 'busy',
+  full: 'full',
+  destroy: 'destroy',
   error: 'error',
   taskError: 'taskError'
 } as const)
@@ -183,9 +184,10 @@ export interface IPool<
    *
    * Events that can currently be listened to:
    *
-   * - `'full'`: Emitted when the pool is dynamic and the number of workers created has reached the maximum size expected.
    * - `'ready'`: Emitted when the number of workers created in the pool has reached the minimum size expected and are ready.
    * - `'busy'`: Emitted when the number of workers created in the pool has reached the maximum size expected and are executing at least one task.
+   * - `'full'`: Emitted when the pool is dynamic and the number of workers created has reached the maximum size expected.
+   * - '`destroy`': Emitted when the pool is destroyed.
    * - `'error'`: Emitted when an uncaught error occurs.
    * - `'taskError'`: Emitted when an error occurs while executing a task.
    */