+++ CHANGELOG.md
@@ -7,12 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `destroy` event to pool API.
+
 ## [2.6.26] - 2023-08-15
 
 ### Added
 
 - Add kill handler to worker options allowing to execute custom code when worker is killed.
-- Add listTaskFunctions() method to pool API.
+- Add `listTaskFunctions()` method to pool API.
 - SMTP server pool example: nodemailer.
 
 ## [2.6.25] - 2023-08-13